requirements.txt
# Google Cloud & Vertex AI dependencies
google-cloud-aiplatform>=1.36.0
google-cloud-storage>=2.10.0
vertexai>=0.1.0

# Data processing and model libraries
transformers>=4.33.0
datasets>=2.14.0
pandas>=2.0.3
numpy>=1.24.0
torch>=2.0.0

# Flask API backend
flask>=2.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
tqdm>=4.66.0
